requirements.txt
numpy>=1.8.2
pandas>=0.13.1
scipy>=0.13.3
matplotlib>=1.3.1
requests>=2.18.4
pillow>=4.3.0
genty>=1.3.2
nose>=1.3.7
rednose>=1.3.0
